@@ -1,13 +1,16 @@
 import React from "react";
-import ReactDOM from "react-dom";
-import App from "./app";
-import { BrowserRouter as Router } from "react-router-dom";
+import { createRoot } from "react-dom/client";
+import { BrowserRouter } from "react-router-dom";
+import { App } from "./app";
+import "./styles/styles.css";
 
-import "./index.css";
+const container = document.getElementById("root");
+const root = createRoot(container);
 
-ReactDOM.render(
-  <Router>
-    <App />
-  </Router>,
-  document.getElementById("root")
+root.render(
+  <React.StrictMode>
+    <BrowserRouter>
+      <App />
+    </BrowserRouter>
+  </React.StrictMode>
 );
